refactor(UserInteractionController): controller name 변경과 InputView와 OutputView 객체로 변경
- InputView와 OutputView를 class -> object로 변경하여, 기존 UserInface에서 property로 있는 거 삭제
- UserInterface에서 UserInteractionController로 class name 변경
